+++ java-service-processor/src/test/resources/definition/TestIdNonStringWithFormat.java
@@ -1,0 +1,14 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+import nbbrd.service.ServiceId;
+
+@ServiceDefinition
+interface TestIdNonStringWithFormat {
+
+    enum Category { A, B }
+
+    @ServiceId(formatMethodName = "name")
+    Category getCategory();
+}
+