+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add custom signature of batch method [#322](https://github.com/nbbrd/java-service-util/issues/322)
 - Add generation of batch provider [#161](https://github.com/nbbrd/java-service-util/issues/161)
 - Add topLevelClassName mustache template variable [#566](https://github.com/nbbrd/java-service-util/issues/566)
+- Add support of types representable as String to ServiceId annotation [#325](https://github.com/nbbrd/java-service-util/issues/325)
 
 ### Changed